⚙️ feat: google go code style compilance and deprecation cleaning
