+++ build.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-go build -o dist/yfs ./binutils/yfs_binutils.go
+go build -o dist/yfs ./cmd/*.go
