+++ build_proto.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 export PATH="$HOME/go/bin:$PATH"
-protoc --go_out=. --go_opt=paths=source_relative ./lib/yfs.proto
+protoc --go_out=. --go_opt=paths=source_relative ./yfs.proto
