+++ go.mod
@@ -1,0 +1,7 @@
+module github.com/sammwyy/yfs
+
+go 1.22
+
+toolchain go1.24.4
+
+require google.golang.org/protobuf v1.36.6
