+++ go.work
@@ -1,4 +1,4 @@
 go 1.24.4
 
-use ./binutils
-use ./lib
+use ./cmd
+use .
