--- lib/go.mod
@@ -1,5 +1,0 @@
-module github.com/sammwyy/yfs/lib
-
-go 1.21
-
-require google.golang.org/protobuf v1.36.6
